requirements-finetuning.txt
# Fine-tuning dependencies - locked versions for Python 3.9 + macOS
# Tested and verified - NO ERRORS

# Core - MUST install first (NumPy 1.x for compatibility)
numpy<2.0,>=1.24.0

# PyTorch
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Transformers ecosystem - compatible versions
transformers==4.40.2
tokenizers==0.19.1
peft==0.10.0
accelerate==0.28.0
datasets==2.18.0
trl==0.8.6

# GGUF conversion for Ollama
llama-cpp-python==0.2.20

# Project dependencies
python-dotenv==1.0.0
pydantic==2.6.4
boto3==1.34.69
botocore==1.34.69

# Utilities
tqdm==4.66.2
psutil==5.9.8

# Additional for stability
safetensors==0.4.2
huggingface-hub==0.21.4

